Move core plugins out of sub folders
