+++ packages/roosterjs-editor-core/lib/corePlugins/CopyPastePlugin.ts
