+++ packages/roosterjs-editor-core/lib/corePlugins/DOMEventPlugin.ts
