+++ packages/roosterjs-editor-core/lib/corePlugins/EntityPlugin.ts
