+++ packages/roosterjs-editor-core/lib/corePlugins/LifecyclePlugin.ts
