+++ packages/roosterjs-editor-core/lib/corePlugins/MouseUpPlugin.ts
