+++ packages/roosterjs-editor-core/lib/corePlugins/PendingFormatStatePlugin.ts
