+++ packages/roosterjs-editor-core/lib/corePlugins/TypeAfterLinkPlugin.ts
