+++ packages/roosterjs-editor-core/lib/corePlugins/TypeInContainerPlugin.ts
